requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
conformer==0.3.2
deepspeed==0.14.2; sys_platform == 'linux'
diffusers==0.32.0
gdown==5.1.0
gradio==4.32.2
grpcio==1.57.0
grpcio-tools==1.57.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
inflect==7.3.1
librosa==0.10.2
celery[redis]
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
passlib
bcrypt==4.0.1
cryptography
psycopg2-binary
python-dotenv 
lightning==2.2.4
matplotlib==3.7.5
networkx==3.1
omegaconf==2.3.0
onnxruntime-gpu==1.16.0; sys_platform == 'linux'
openai-whisper==20231117
protobuf==4.25
pydantic==2.7.0
rich==13.7.1
soundfile==0.12.1
tensorboard==2.14.0
torch==2.3.1
torchaudio==2.3.1
wget==3.2
fastapi==0.111.0
fastapi-cli==0.0.4
WeTextProcessing==1.0.3
opencc-python-reimplemented
g2pw
pyarrow
datasets
https://www.modelscope.cn/models/speech_tts/speech_kantts_ttsfrd/resolve/master/ttsfrd_dependency-0.1-py3-none-any.whl
https://www.modelscope.cn/models/speech_tts/speech_kantts_ttsfrd/resolve/master/ttsfrd-0.3.9-cp310-cp310-linux_x86_64.whl
